Move all database related files to new db directory
