+++ db/seeders/.empty
@@ -1,0 +1,1 @@
+Delete once we have seeds
